// engonow-library/interfaces.ts
export enum UserRole {
    ADMIN = "Quản trị viên",
    SYS_ADMIN = "Quản trị viên hệ thống",
    EXAM_ADMIN = "Quản trị viên đề thi",
    STUDENT = "Học viên",
    PAID_USER = "Người dùng có trả phí",
    NONPAID_USER = "Người dùng không trả phí"
}
// ...
export interface UpdateAccountPayload {
    username?: string;
    password?: string;
    firstName?: string;
    lastName?: string;
    phoneNumber?: string;
    dateOfBirth?: Date;
    roles?: UserRole[]; // Just ADMIN, SYS_ADMIN allowed
}

export interface UpdateAvatarPayload {
    avatar: File;
}